+++ SwiftyButton/FlatButton.swift
@@ -14,6 +14,7 @@ public class FlatButton: UIButton {
     public enum Defaults {
         public static var color = UIColor(red: 52 / 255, green: 152 / 255, blue: 219 / 255, alpha: 1)
         public static var highlightedColor =  UIColor(red: 41 / 255, green: 128 / 255, blue: 185 / 255, alpha: 1)
+        public static var selectedColor =  UIColor(red: 41 / 255, green: 128 / 255, blue: 185 / 255, alpha: 1)
         public static var disabledColor = UIColor(red: 149 / 255, green: 165 / 255, blue: 166 / 255, alpha: 1)
         public static var cornerRadius: CGFloat = 3
     }
@@ -27,6 +28,13 @@ public class FlatButton: UIButton {
     
     @IBInspectable
     public var highlightedColor: UIColor = Defaults.highlightedColor {
+        didSet {
+            updateBackgroundImages()
+        }
+    }
+    
+    @IBInspectable
+    public var selectedColor: UIColor = Defaults.selectedColor {
         didSet {
             updateBackgroundImages()
         }
@@ -71,10 +79,12 @@ public class FlatButton: UIButton {
         
         let normalImage = Utils.buttonImage(color: color, shadowHeight: 0, shadowColor: .clear, cornerRadius: cornerRadius)
         let highlightedImage = Utils.highlightedButtonImage(color: highlightedColor, shadowHeight: 0, shadowColor: .clear, cornerRadius: cornerRadius, buttonPressDepth: 0)
+        let selectedImage = Utils.buttonImage(color: selectedColor, shadowHeight: 0, shadowColor: .clear, cornerRadius: cornerRadius)
         let disabledImage = Utils.buttonImage(color: disabledColor, shadowHeight: 0, shadowColor: .clear, cornerRadius: cornerRadius)
         
         setBackgroundImage(normalImage, for: .normal)
         setBackgroundImage(highlightedImage, for: .highlighted)
+        setBackgroundImage(selectedImage, for: .selected)
         setBackgroundImage(disabledImage, for: .disabled)
     }
 }