+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Change Log
 
 ## [master]
 
+ - Added a `FlatButton.selectedColor` property to customize the appearance of a flat button in its selected state.
  - Adding Carthage support.
 
 ## [0.6.0]